Starting to recieve initial data

+++ life/src/main.rs
@@ -89,17 +89,17 @@ fn main() {
     let mut fromup:Vec<i32>;  //arrays to send and to receive
     for g in 1..generations{ //generations for loop
         if rank!=size-1 {// all except for last send down
-          world.process_at_rank(rank+1).send(&slice[s-1][..]);
+          world.process_at_rank(rank+1).send(&slice[(s-1) as usize][..]);
         } else {
-          fromdown = vec![0; n]; // last one generates empty stripe "from down"
+          fromdown = vec![0; n as usize]; // last one generates empty stripe "from down"
         }
         if rank!=0{ // all except for first receive from up
             let (fromup, status) = world.process_at_rank(rank-1).receive_vec::<i32>();
         } else {
-            fromup = vec![0; n]; // first one generats empty line "from up"
+            fromup = vec![0; n as usize]; // first one generats empty line "from up"
         }
         if rank!=0{ // all except for first send up
-            world.process_at_rank(rank-1).send(&slice[s-1][..]);
+            world.process_at_rank(rank-1).send(&slice[(s-1) as usize][..]);
         }
         if rank!=size-1 { // all except for last receive from down
             let (fromdown, status) = world.process_at_rank(rank+1).receive_vec::<i32>();
